requirements.txt
django==2.2.13
djangorestframework==3.10.2
djangorestframework-simplejwt==4.4.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
psycopg2==2.8
Pillow==7.1.1
whitenoise==5.1.0
dj_database_url==0.5.0
requests==2.22.0
django-cors-headers==3.2.0
